+++ lib/pages/home_page/home_page.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:sample/pages/home_page/widgets/header_ribbon.dart';
+
+class MyHomePage extends StatelessWidget {
+  const MyHomePage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: HeaderRibbon(),
+    );
+  }
+}
+
+
